+++ crates/rtlola-parser/src/ast.rs
@@ -433,9 +433,9 @@ pub enum StreamAccessKind {
     /// Hold access for *incompatible* stream types, returns previous known value
     Hold,
     /// Optional access, returns value if it exists, called by `.get()`
-    Optional,
+    Get,
     /// Boolean Typed access, returning true if the target stream received a new value at the current timestamp. Called with `.is_fresh()`.
-    UpdateCheck,
+    Fresh,
 }
 
 /// Describes the operation used to access a stream with a offset