+++ crates/rtlola-parser/src/ast/print.rs
@@ -184,8 +184,8 @@ impl Display for Expression {
                 match access {
                     StreamAccessKind::Sync => write!(f, "{}", expr),
                     StreamAccessKind::Hold => write!(f, "{}.hold()", expr),
-                    StreamAccessKind::Optional => write!(f, "{}.get()", expr),
-                    StreamAccessKind::UpdateCheck => write!(f, "{}.is_fresh()", expr),
+                    StreamAccessKind::Get => write!(f, "{}.get()", expr),
+                    StreamAccessKind::Fresh => write!(f, "{}.is_fresh()", expr),
                 }
             },
             ExpressionKind::Default(expr, val) => write!(f, "{}.defaults(to: {})", expr, val),